[sdarq][backend] Replace firebase_admin with google-cloud-firestore

+++ sdarq/backend/src/app.py
@@ -5,8 +5,6 @@ from flask_api import FlaskAPI
 from flask_cors import CORS, cross_origin
 from google.cloud import bigquery
 from google.cloud import pubsub_v1
-from firebase_admin import credentials
-from firebase_admin import firestore
 from jira import JIRA
 import slacknotify as slacknotify
 import defectdojo as wrapper
@@ -14,8 +12,9 @@ import os
 import re
 import slack
 import datetime
-import firebase_admin
 import json
+from google.cloud import resource_manager
+from google.cloud import firestore
 
 # Env variables
 dojo_host = os.getenv('dojo_host')
@@ -26,7 +25,6 @@ jira_username = os.getenv('jira_username')
 jira_api_token = os.getenv('jira_api_token')
 jira_instance = os.getenv('jira_instance')
 sdarq_host = os.getenv('sdarq_host')
-firebase_project_id = os.getenv('firebase_project_id')
 
 # Instantiate the DefectDojo backend wrapper
 dd = wrapper.DefectDojoAPI(dojo_host, dojo_api_key, dojo_user, debug=True)
@@ -35,14 +33,6 @@ app = FlaskAPI(__name__)
 # Instantiate the Jira backend wrapper
 global jira
 jira = JIRA(basic_auth=(jira_username, jira_api_token), options={'server': jira_instance })
-
-
-# Instantiate Firebase for Firestore
-cred = credentials.ApplicationDefault()
-firebase_admin.initialize_app(cred, {
-    'projectId': firebase_project_id,
-})
-db = firestore.client()
 
 
 @app.route('/health/', methods=['GET'])
@@ -185,7 +175,7 @@ def cis_results(project_id):
 
 
 @app.route('/cis_scan/', methods=['POST'])
-@cross_origin(origins=sdarq_host)
+@cross_origin(origins=sdarq_host')
 def cis_scan():
     """
     Scans a specific google project
@@ -199,39 +189,46 @@ def cis_scan():
     message = ""
     message = message.encode("utf-8")
     firestore_collection = 'cis_scans'
+    all_projects = []
 
     if re.match(pattern, user_project_id):
-        publisher = pubsub_v1.PublisherClient()
-        topic_path = publisher.topic_path(project_id, topic_name)
-        if 'slack_channel' in json_data:
-            slack_channel = json_data['slack_channel']
-            publisher.publish(topic_path,
-                              data=message,
-                              GCP_PROJECT_ID=user_project_id,
-                              SLACK_CHANNEL=slack_channel,
-                              RESULTS_URL=results_url,
-                              FIRESTORE_COLLECTION=firestore_collection)
-        else:
-            publisher.publish(topic_path,
-                              data=message,
-                              GCP_PROJECT_ID=user_project_id,
-                              FIRESTORE_COLLECTION=firestore_collection)
+        for p in resource_manager.Client().list_projects():
+            all_projects.append(p.name)
+            if user_project_id in all_projects:
+                publisher = pubsub_v1.PublisherClient()
+                topic_path = publisher.topic_path(project_id, topic_name)
+                if 'slack_channel' in json_data:
+                    slack_channel = json_data['slack_channel']
+                    publisher.publish(topic_path,
+                                    data=message,
+                                    GCP_PROJECT_ID=user_project_id,
+                                    SLACK_CHANNEL=slack_channel,
+                                    RESULTS_URL=results_url,
+                                    FIRESTORE_COLLECTION=firestore_collection)
+                else:
+                    publisher.publish(topic_path,
+                                    data=message,
+                                    GCP_PROJECT_ID=user_project_id,
+                                    FIRESTORE_COLLECTION=firestore_collection)
 
-        # Check if a document exists in Firestore
-        check = False
-        while check is False:
-            doc_ref = db.collection(
-                firestore_collection).document(user_project_id)
-            doc = doc_ref.get()
-            if doc.exists:
-                check = True
-            else:
+                # Check if a document exists in Firestore
                 check = False
+                while check is False:
+                    db = firestore.Client()
+                    doc_ref = db.collection(
+                        firestore_collection).document(user_project_id)
+                    doc = doc_ref.get()
+                    if doc.exists:
+                        check = True
+                    else:
+                        check = False
 
-        db.collection(firestore_collection).document(
-            user_project_id).delete()
+                db.collection(firestore_collection).document(
+                    user_project_id).delete()
 
-        return Response(json.dumps({'statusText': 'Doc found!', 'status': 'true'}), status=200, mimetype='application/json')
+                return Response(json.dumps({'statusText': 'Doc found!', 'status': 'true'}), status=200, mimetype='application/json')
+        else:
+            return Response(json.dumps({'statusText': ' Project does not exist! Please make sure this project is part of the organization. ONLY projects that are part of BROAD INSTITUTE organization can be scanned from CIS tool. ', 'status': 'false', 'statusCode': '404'}), status=200, mimetype='application/json')
 
 
 if __name__ == "__main__":
